pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pytest-aimrt"
version = "0.1.0"
description = "A pytest-based testing framework for AimRT examples"
readme = "README.md"
license = {text = "Apache-2.0"}
authors = [
    {name = "AimRT Team"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Testing",
    "Framework :: Pytest"
]
requires-python = ">=3.8"
dependencies = [
    "pytest>=6.0.0",
    "pytest-xdist>=2.0.0",
    "pyyaml>=5.4.0",
    "psutil>=5.8.0",
    "dataclasses; python_version<'3.7'"
]

[project.optional-dependencies]
dev = [
    "pytest-cov>=2.12.0",
    "black>=21.0.0",
    "flake8>=3.9.0",
    "mypy>=0.910",
    "pre-commit>=2.15.0"
]

[project.urls]
Homepage = "https://github.com/AimRT/AimRT"
Repository = "https://github.com/AimRT/AimRT"
Documentation = "https://aimrt.org"

[project.entry-points.pytest11]
pytest_aimrt = "pytest_aimrt.fixtures"

[tool.setuptools.packages.find]
include = ["pytest_aimrt*"]

[tool.setuptools.package-data]
pytest_aimrt = ["py.typed"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers"
testpaths = [
    "tests",
    "pytest_aimrt/examples"
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "aimrt: marks tests as AimRT framework tests"
]